# services/server/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hot Cross Buns Sync Server API",
    "description": "REST API for the Hot Cross Buns sync server. All endpoints except /health and /api/v1/auth/* require a Bearer JWT token when AUTH_REQUIRED=true.",
    "version": "1.0.0",
    "contact": {
      "name": "Hot Cross Buns"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT with sub=userID, exp=30 days. Not required when AUTH_REQUIRED=false."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "One or more fields are invalid."
              },
              "details": {
                "type": "array",
                "items": { "type": "string" },
                "example": []
              }
            },
            "required": ["code", "message", "details"]
          }
        },
        "required": ["error"]
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "urgent" },
          "color": { "type": "string", "nullable": true, "example": "#f38ba8" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "userId", "name", "createdAt"]
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "listId": { "type": "string", "format": "uuid" },
          "parentTaskId": { "type": "string", "format": "uuid", "nullable": true },
          "title": { "type": "string", "example": "Buy groceries" },
          "content": { "type": "string", "nullable": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 3, "description": "0=none, 1=low, 2=medium, 3=high" },
          "status": { "type": "integer", "minimum": 0, "maximum": 1, "description": "0=open, 1=completed" },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true },
          "dueTimezone": { "type": "string", "nullable": true, "example": "Asia/Singapore" },
          "recurrenceRule": { "type": "string", "nullable": true, "description": "RFC 5545 RRULE string" },
          "sortOrder": { "type": "integer" },
          "completedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true },
          "subtasks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Task" }
          },
          "tags": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Tag" }
          }
        },
        "required": ["id", "userId", "listId", "title", "priority", "status", "sortOrder", "createdAt", "updatedAt"]
      },
      "CreateTaskRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 500, "example": "Buy groceries" },
          "content": { "type": "string", "nullable": true, "maxLength": 10000 },
          "priority": { "type": "integer", "minimum": 0, "maximum": 3 },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true },
          "dueTimezone": { "type": "string", "nullable": true },
          "parentTaskId": { "type": "string", "format": "uuid", "nullable": true },
          "recurrenceRule": { "type": "string", "nullable": true }
        },
        "required": ["title"]
      },
      "UpdateTaskRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 500 },
          "content": { "type": "string", "nullable": true },
          "listId": { "type": "string", "format": "uuid" },
          "parentTaskId": { "type": "string", "format": "uuid", "nullable": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 3 },
          "status": { "type": "integer", "minimum": 0, "maximum": 1 },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true },
          "dueTimezone": { "type": "string", "nullable": true },
          "recurrenceRule": { "type": "string", "nullable": true },
          "sortOrder": { "type": "integer" }
        }
      },
      "MoveTaskRequest": {
        "type": "object",
        "properties": {
          "listId": { "type": "string", "format": "uuid" },
          "sortOrder": { "type": "integer" }
        },
        "required": ["listId"]
      },
      "CompleteTaskResponse": {
        "type": "object",
        "properties": {
          "completed": { "$ref": "#/components/schemas/Task" },
          "next": {
            "nullable": true,
            "allOf": [{ "$ref": "#/components/schemas/Task" }],
            "description": "Next occurrence if the task is recurring, null otherwise."
          }
        },
        "required": ["completed", "next"]
      },
...
    }
  },
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/v1/lists/{listId}/tasks": {
      "parameters": [
        {
          "name": "listId",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "Parent list UUID"
        }
      ],
      "post": {
        "tags": ["tasks"],
        "summary": "Create a task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTaskRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Validation error or nesting depth exceeded.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Parent task not found (when parentTaskId is provided).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["tasks"],
        "summary": "Get tasks by list",
        "description": "Returns tasks for a list with nested subtasks and tags.",
        "parameters": [
          {
            "name": "includeCompleted",
            "in": "query",
            "schema": { "type": "boolean", "default": false },
            "description": "Include completed tasks in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Array of tasks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          },
          "400": {
            "description": "Invalid list ID.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "Task UUID"
        }
      ],
      "get": {
        "tags": ["tasks"],
        "summary": "Get a task by ID",
        "responses": {
          "200": {
            "description": "Task found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Invalid task ID.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Task not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["tasks"],
        "summary": "Update a task",
        "description": "Partial update. Only include fields to change.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Task not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["tasks"],
        "summary": "Delete a task (soft delete)",
        "responses": {
          "204": {
            "description": "Task deleted."
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Task not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/move": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "Task UUID"
        }
      ],
      "post": {
        "tags": ["tasks"],
        "summary": "Move a task to another list",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MoveTaskRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task moved. Returns the full updated task.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Validation error (missing listId).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Task not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/complete": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "Task UUID"
        }
      ],
      "post": {
        "tags": ["tasks"],
        "summary": "Complete a task",
        "description": "Marks the task as completed. If the task has a recurrence rule, creates the next occurrence and returns both.",
        "responses": {
          "200": {
            "description": "Task completed. next is non-null for recurring tasks.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompleteTaskResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Task not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}